+++ CHANGELOG.md
@@ -1,3 +1,3 @@
 ## Changelog
-- update classtweaker version `0.2` -> `0.3.0-beta.2`
-- update NeoForge version `26.1.1.10-beta` -> `26.1.2.22-beta`
+- update NeoForge version `26.1.2.22-beta` -> `26.2.0.8-beta`
+- fix when modifying the accessibility of fields in the `Biome` class that was incompatible with Neoforge's built-in `ReplaceFieldWithGetterAccess` class-processor ([#1](https://github.com/CagayakeGirls/FoxifiedClassTweaker/issues/1))